requirements.txt
tabulate
numpy
pandas
scipy
pyyaml
torch
torchvision
